+++ CHANGES.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Changes in 2.x (date TBD)
+-------------------------
+
+Added support for multiple databases:
+
+- Fixed cache overlap issue when using multiple database: cache key now includes the database alias.
+- Now properly support multiple databases and translatable models duplication (fully documented on page `Duplicating instances, using multiple databases and more...`).
+- Now clearly identify some marginal case django-parler does not support, and raise an exception instead of silently failing and possibly corrupting the database.
+
+
 Changes in 2.3 (2021-11-18)
 ---------------------------